requirement.txt
selenium
pytest
python-dotenv
pyyaml
faker
